# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the RTO Standards 2025 Practice Guide
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProgressStatus =
  (typeof ProgressStatus)[keyof typeof ProgressStatus];

export const ProgressStatus = {
  not_started: "not_started",
  planning: "planning",
  in_progress: "in_progress",
  embedded: "embedded",
} as const;

export type StrategyCategory =
  (typeof StrategyCategory)[keyof typeof StrategyCategory];

export const StrategyCategory = {
  classroom: "classroom",
  assessment: "assessment",
  learner_support: "learner_support",
  workforce: "workforce",
  governance: "governance",
  compliance: "compliance",
  reflection: "reflection",
} as const;

export type StrategyEffort =
  (typeof StrategyEffort)[keyof typeof StrategyEffort];

export const StrategyEffort = {
  quick_win: "quick_win",
  ongoing: "ongoing",
  deep_change: "deep_change",
} as const;

export interface QualityAreaSummary {
  id: number;
  code: string;
  title: string;
  tagline: string;
  description: string;
  standardCount: number;
  embeddedCount: number;
  inProgressCount: number;
  planningCount: number;
  notStartedCount: number;
}

export interface StandardSummary {
  id: number;
  qualityAreaId: number;
  qualityAreaCode: string;
  qualityAreaTitle: string;
  code: string;
  title: string;
  intent: string;
  strategyCount: number;
  status: ProgressStatus;
}

export interface QualityAreaDetail {
  id: number;
  code: string;
  title: string;
  tagline: string;
  description: string;
  standards: StandardSummary[];
}

export interface Strategy {
  id: number;
  standardId: number;
  standardCode: string;
  standardTitle: string;
  qualityAreaId: number;
  qualityAreaCode: string;
  title: string;
  summary: string;
  steps: string[];
  category: StrategyCategory;
  effort: StrategyEffort;
  timeEstimate: string;
  favorited: boolean;
}

export interface ProgressEntry {
  standardId: number;
  status: ProgressStatus;
  updatedAt: string;
}

export interface StandardDetail {
  id: number;
  qualityAreaId: number;
  qualityAreaCode: string;
  qualityAreaTitle: string;
  code: string;
  title: string;
  intent: string;
  whatItMeans: string;
  keyPractices: string[];
  evidenceExamples: string[];
  strategies: Strategy[];
  progress?: ProgressEntry;
}

export interface FeaturedStrategy {
  strategy: Strategy;
  rationale: string;
}

export interface SetProgressBody {
  status: ProgressStatus;
}

export interface Note {
  id: number;
  standardId: number;
  standardCode: string;
  standardTitle: string;
  body: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  standardId: number;
  /**
   * @minLength 1
   * @maxLength 10000
   */
  body: string;
}

export interface UpdateNoteBody {
  /**
   * @minLength 1
   * @maxLength 10000
   */
  body: string;
}

export interface Favorite {
  strategyId: number;
  strategy: Strategy;
  createdAt: string;
}

export interface AddFavoriteBody {
  strategyId: number;
}

export interface QualityAreaProgress {
  qualityAreaId: number;
  qualityAreaCode: string;
  qualityAreaTitle: string;
  standardCount: number;
  embeddedCount: number;
  inProgressCount: number;
  planningCount: number;
  notStartedCount: number;
  percentComplete: number;
}

export interface DashboardSummary {
  totalStandards: number;
  totalStrategies: number;
  embeddedCount: number;
  inProgressCount: number;
  planningCount: number;
  notStartedCount: number;
  overallPercentComplete: number;
  favoriteCount: number;
  noteCount: number;
  qualityAreaProgress: QualityAreaProgress[];
  recentNotes: Note[];
  recentlyUpdatedStandards: StandardSummary[];
  featured: FeaturedStrategy;
}

export type ListStandardsParams = {
  qualityAreaId?: number;
  search?: string;
};

export type ListStrategiesParams = {
  standardId?: number;
  category?: StrategyCategory;
  effort?: StrategyEffort;
};

export type ListNotesParams = {
  standardId?: number;
};